% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vif.R
\name{calculate_vif}
\alias{calculate_vif}
\title{Variance Inflation Factor (VIF) Calculation}
\usage{
calculate_vif(model)
}
\arguments{
\item{model}{A fitted linear model (object of class \code{lm}).}
}
\value{
A named vector of VIF values for each predictor variable.
}
\description{
This function calculates the VIF for each predictor variable in a linear model.
}
\examples{
model <- lm(mpg ~ disp + hp + wt, data = mtcars)
calculate_vif(model)
}
