% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_outliers.R
\name{iqr_outliers}
\alias{iqr_outliers}
\title{IQR Method for Detecting Outliers}
\usage{
iqr_outliers(data)
}
\arguments{
\item{data}{A numeric vector containing the data.}
}
\value{
A list containing the outliers and the cleaned data.
}
\description{
The method detects outliers by identifying values that are below Q1 - 1.5 * IQR or above Q3 + 1.5 * IQR.
}
\details{
This function identifies outliers based on the Interquartile Range (IQR) method.
Outliers are defined as values outside the range calculated using Q1 and Q3 with a multiplier of 1.5.
}
\examples{
iqr_outliers(c(10, 12, 15, 100, 13))
}
