% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_imputation.R
\name{mean_imputation}
\alias{mean_imputation}
\title{Mean Imputation}
\usage{
mean_imputation(data, col_name)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{col_name}{The name of the column (as a string) to perform mean imputation on.}
}
\value{
A data frame with missing values in the specified column replaced by the mean.
}
\description{
This function replaces missing values in a column with the mean of that column.
}
\examples{
data <- data.frame(x = c(1, 2, NA, 4, 5))
mean_imputation(data, "x")
}
