% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot_encode.R
\name{one_hot_encode}
\alias{one_hot_encode}
\title{One-Hot Encoding Function}
\usage{
one_hot_encode(data, col_name)
}
\arguments{
\item{data}{A data frame containing the categorical variable.}

\item{col_name}{The name of the column (as a string) to perform one-hot encoding on.}
}
\value{
A data frame with one-hot encoded columns replacing the original column.
}
\description{
This function performs one-hot encoding on a categorical vector, converting it to binary columns.
}
\examples{
one_hot_encode(data.frame(category = c("A", "B", "A", "C")), "category")
}
