% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove Duplicates}
\usage{
remove_duplicates(data)
}
\arguments{
\item{data}{A data frame to remove duplicates from.}
}
\value{
A data frame with duplicates removed.
}
\description{
This function removes duplicate rows from a data frame.
}
\examples{
remove_duplicates(data.frame(a = c(1, 2, 2), b = c(3, 4, 4)))
}
