% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retain_duplicates.R
\name{retain_duplicates}
\alias{retain_duplicates}
\title{Retain Duplicates}
\usage{
retain_duplicates(data)
}
\arguments{
\item{data}{A data frame to extract duplicates from.}
}
\value{
A data frame containing only duplicate rows.
}
\description{
This function retains only the duplicate rows in a data frame.
}
\examples{
retain_duplicates(data.frame(a = c(1, 2, 2), b = c(3, 4, 4)))
}
