% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardization Function}
\usage{
standardize(data)
}
\arguments{
\item{data}{A numeric vector to be standardized.}
}
\value{
A numeric vector with standardized values.
}
\description{
This function standardizes a numeric vector using the formula:
z = (x - mean(x)) / sd(x)
}
\examples{
standardize(c(10, 20, 30, 40, 50))
}
