% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_loss.R
\name{weighted_loss}
\alias{weighted_loss}
\title{Weighted Loss Function}
\usage{
weighted_loss(y_true, y_pred, class_weights)
}
\arguments{
\item{y_true}{A vector of true class labels.}

\item{y_pred}{A vector of predicted probabilities for the positive class.}

\item{class_weights}{A vector of weights for each class.}
}
\value{
The weighted loss value.
}
\description{
This function calculates a weighted loss for a classification problem.
}
\examples{
weighted_loss(c(1, 0, 1), c(0.9, 0.2, 0.7), c(0.3, 0.7))
}
